requirements.txt
torch
torchaudio
transformers
numpy
pandas
scikit-learn
tqdm
pyyaml
